Polish implementation and documentation
