+++ Sources/Player/Player+Items.swift
@@ -141,22 +141,6 @@ public extension Player {
     }
 }
 
-extension Player {
-    static func items(before index: Int?, in items: Deque<PlayerItem>) -> [PlayerItem] {
-        guard let index else { return [] }
-        let previousIndex = items.index(before: index)
-        guard previousIndex >= 0 else { return [] }
-        return Array(items.suffix(from: previousIndex))
-    }
-
-    static func items(after index: Int?, in items: Deque<PlayerItem>) -> [PlayerItem] {
-        guard let index else { return [] }
-        let nextIndex = items.index(after: index)
-        guard nextIndex < items.count else { return [] }
-        return Array(items.suffix(from: nextIndex))
-    }
-}
-
 private extension Player {
     func canInsert(_ item: PlayerItem, before beforeItem: PlayerItem?) -> Bool {
         guard let beforeItem else { return true }
