+++ Sources/Player/Player+Navigation.swift
@@ -11,7 +11,7 @@ public extension Player {
     /// Check whether returning to the previous item in the deque is possible.`
     /// - Returns: `true` if possible.
     func canReturnToPreviousItem() -> Bool {
-        Self.canReturnToItem(before: currentIndex, in: storedItems)
+        canReturnToItem(before: currentIndex, in: storedItems)
     }
 
     /// Return to the previous item in the deque. Skips failed items.
@@ -31,38 +31,9 @@ public extension Player {
         guard canAdvanceToNextItem() else { return }
         queuePlayer.replaceItems(with: AVPlayerItem.playerItems(from: advancingItems))
     }
-}
 
-public extension Player {
-    /// Check whether returning to the previous content is possible.`
-    /// - Returns: `true` if possible.
-    func canReturnToPrevious() -> Bool {
-        canReturn(before: currentIndex, in: storedItems, streamType: streamType)
-    }
-
-    /// Return to the previous content.
-    func returnToPrevious() {
-        if shouldSeekToStartTime() {
-            seek(near(.zero))
-        }
-        else {
-            returnToPreviousItem()
-        }
-    }
-
-    /// Check whether moving to the next content is possible.`
-    /// - Returns: `true` if possible.
-    func canAdvanceToNext() -> Bool {
-        canAdvanceToNextItem()
-    }
-
-    /// Move to the next content.
-    func advanceToNext() {
-        advanceToNextItem()
-    }
-
-    /// Set the index of the current item.
-    /// - Parameter index: The index to set.
+    /// Set the item at the specified index to become the current one.
+    /// - Parameter index: The item index.
     func setCurrentIndex(_ index: Int) throws {
         guard index != currentIndex else { return }
         guard (0..<storedItems.count).contains(index) else { throw PlaybackError.itemOutOfBounds }
@@ -72,13 +43,8 @@ public extension Player {
 }
 
 extension Player {
-    func canReturn(before index: Int?, in items: Deque<PlayerItem>, streamType: StreamType) -> Bool {
-        if configuration.isSmartNavigationEnabled && streamType == .onDemand {
-            return true
-        }
-        else {
-            return Self.canReturnToItem(before: index, in: items)
-        }
+    func canReturnToItem(before index: Int?, in items: Deque<PlayerItem>) -> Bool {
+        !Self.items(before: index, in: items).isEmpty
     }
 
     func canAdvanceToItem(after index: Int?, in items: Deque<PlayerItem>) -> Bool {
@@ -87,8 +53,6 @@ extension Player {
 }
 
 private extension Player {
-    static let startTimeThreshold: TimeInterval = 3
-
     /// Return the list of items to be loaded to return to the previous (playable) item.
     var returningItems: [PlayerItem] {
         Self.items(before: currentIndex, in: storedItems)
@@ -99,16 +63,17 @@ private extension Player {
         Self.items(after: currentIndex, in: storedItems)
     }
 
-    static func canReturnToItem(before index: Int?, in items: Deque<PlayerItem>) -> Bool {
-        !Self.items(before: index, in: items).isEmpty
+    static func items(before index: Int?, in items: Deque<PlayerItem>) -> [PlayerItem] {
+        guard let index else { return [] }
+        let previousIndex = items.index(before: index)
+        guard previousIndex >= 0 else { return [] }
+        return Array(items.suffix(from: previousIndex))
     }
 
-    private func isFarFromStartTime() -> Bool {
-        time.isValid && timeRange.isValid && (time - timeRange.start).seconds >= Self.startTimeThreshold
-    }
-
-    private func shouldSeekToStartTime() -> Bool {
-        guard configuration.isSmartNavigationEnabled else { return false }
-        return (streamType == .onDemand && isFarFromStartTime()) || !canReturnToPreviousItem()
+    static func items(after index: Int?, in items: Deque<PlayerItem>) -> [PlayerItem] {
+        guard let index else { return [] }
+        let nextIndex = items.index(after: index)
+        guard nextIndex < items.count else { return [] }
+        return Array(items.suffix(from: nextIndex))
     }
 }
