+++ Sources/Player/Player+PlaybackSpeed.swift
@@ -27,20 +27,6 @@ public extension Player {
 }
 
 extension Player {
-    private static func playbackSpeedRange(for timeRange: CMTimeRange, itemDuration: CMTime, time: CMTime) -> ClosedRange<Float>? {
-        let streamType = StreamType(for: timeRange, itemDuration: itemDuration)
-        switch streamType {
-        case .live:
-            return 1...1
-        case .dvr where time > timeRange.end - CMTime(value: 5, timescale: 1):
-            return 0.1...1
-        case .unknown:
-            return nil
-        default:
-            return 0.1...2
-        }
-    }
-
     func playbackSpeedUpdatePublisher() -> AnyPublisher<PlaybackSpeedUpdate, Never> {
         Publishers.Merge3(
             desiredPlaybackSpeedUpdatePublisher(),
@@ -56,8 +42,24 @@ extension Player {
             .map { .value($0) }
             .eraseToAnyPublisher()
     }
+}
 
-    private func supportedPlaybackSpeedPublisher() -> AnyPublisher<PlaybackSpeedUpdate, Never> {
+private extension Player {
+    static func playbackSpeedRange(for timeRange: CMTimeRange, itemDuration: CMTime, time: CMTime) -> ClosedRange<Float>? {
+        let streamType = StreamType(for: timeRange, itemDuration: itemDuration)
+        switch streamType {
+        case .live:
+            return 1...1
+        case .dvr where time > timeRange.end - CMTime(value: 5, timescale: 1):
+            return 0.1...1
+        case .unknown:
+            return nil
+        default:
+            return 0.1...2
+        }
+    }
+
+    func supportedPlaybackSpeedPublisher() -> AnyPublisher<PlaybackSpeedUpdate, Never> {
         queuePlayer.publisher(for: \.currentItem)
             .weakCapture(self)
             .map { item, player -> AnyPublisher<PlaybackSpeedUpdate, Never> in
@@ -85,7 +87,7 @@ extension Player {
 
     // Publish speed updates triggered from `AVPlayerViewController`. Not necessary on tvOS since the standard UI
     // does not provide speed controls.
-    private func avPlayerViewControllerPlaybackSpeedUpdatePublisher() -> AnyPublisher<PlaybackSpeedUpdate, Never> {
+    func avPlayerViewControllerPlaybackSpeedUpdatePublisher() -> AnyPublisher<PlaybackSpeedUpdate, Never> {
 #if os(iOS)
         queuePlayer.publisher(for: \.rate)
             .filter { rate in
